+++ presentation/src/main/kotlin/ru/endlesscode/miet/orioks/presentation/subjects/fragment/SubjectsFragment.kt
@@ -10,8 +10,7 @@ import ru.endlesscode.miet.orioks.DummyData
 import ru.endlesscode.miet.orioks.R
 import ru.endlesscode.miet.orioks.internal.Screens
 import ru.endlesscode.miet.orioks.internal.di.DI
-import ru.endlesscode.miet.orioks.model.entity.Subject
-import ru.endlesscode.miet.orioks.model.entity.SubjectForm
+import ru.endlesscode.miet.orioks.model.Subject
 import ru.endlesscode.miet.orioks.presentation.common.fragment.BaseFragment
 import ru.endlesscode.miet.orioks.presentation.subjects.adapter.SubjectsAdapter
 import ru.terrakok.cicerone.Router
